화면 디자인 다시 수정
