+++ src/main/resources/templates/book/bookList.html
@@ -12,6 +12,7 @@
             max-width: 960px;
             margin: 0 auto;
             padding: 20px;
+            font-family: 'Nanum Gothic', sans-serif;
         }
 
         .top-list {
@@ -72,10 +73,6 @@
 
         .book-title {
             font-weight: bold;
-            margin-top: 10px;
-        }
-
-        .book-price {
             margin-top: 5px;
         }
 
@@ -103,13 +100,13 @@
         }
 
         .card-img-top {
+            justify-content: center;
             flex-shrink: 0;
             width: 150px;
             margin-right: 20px;
         }
 
         .card-title {
-            font-family: 'Nanum Gothic', sans-serif;
             font-size: 16px;
             font-weight: bold;
             margin-bottom: 10px;
