+++ src/main/resources/templates/book/orderSuccess.html
@@ -1,10 +1,24 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      layout:decorate="~{layout/layout}">
 <head>
   <meta charset="UTF-8">
   <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" rel="stylesheet">
   <title>Order Success</title>
   <style>
+        .container {
+            max-width: 960px;
+            margin: 0 auto;
+            padding: 20px;
+        }
+
+        body {
+            font-family: 'Nanum Gothic', sans-serif;
+            background-color: #ffffff;
+            margin: 0;
+            padding: 0;
+        }
        table {
             border-collapse: collapse;
             width: 100%;
@@ -14,51 +28,81 @@
             text-align: left;
             border-bottom: 1px solid #ddd;
         }
+        .container {
+            background-color: #b2d8e5;
+            padding: 20px;
+            border: 3px solid #b2d8e5;
+            border-radius: 5px;
+        }
+        .order-info {
+            background-color: #fff;
+            padding: 20px;
+            border: 3px solid #b2d8e5;
+            border-radius: 5px;
+            margin-top: 20px;
+        }
+        .order-type {
+            background-color: #fff;
+            padding: 10px;
+            border: 3px solid #b2d8e5;
+            border-radius: 5px;
+            margin-top: 20px;
+        }
+        h1 {
+            font-size: 24px;
+            font-weight: bold;
+            margin-bottom: 20px;
+        }
+        h2 {
+            font-size: 20px;
+            font-weight: bold;
+        }
   </style>
 </head>
 <body>
-<div class="container">
-  <h1>주문에 성공하였습니다. 주문 정보를 확인해주세요.</h1>
-  <div th:if="${order}">
-    <h2>주문 정보</h2>
-    <h3>주문한 책 정보:</h3>
-    <table class="table">
-      <thead>
-      <tr>
-        <th></th>
-        <th>제목</th>
-        <th>수량</th>
-        <th>한 권당 가격</th>
-      </tr>
-      </thead>
-      <tbody>
-      <tr th:each="orderItem : ${order.orderItems}">
-        <td>
-          <img th:src="${orderItem.book.imageUrl}" alt="책 이미지" width="100" height="150" />
-        </td>
-        <td th:text="${orderItem.book.title}"></td>
-        <td th:text="${orderItem.quantity}"></td>
-        <td th:text="${orderItem.book.price}"></td>
-      </tr>
-      </tbody>
-    </table>
-    <p>주문번호: <span th:text="${order.orderId}"></span></p>
-    <p>주문 날짜: <span th:text="${order.orderDate}"></span></p>
-    <p>이름: <span th:text="${order.name}"></span></p>
-    <p>전화번호: <span th:text="${order.phoneNumber}"></span></p>
-    <p>주소: <span th:text="${order.address.address1}"></span>
-      <span th:text="${order.address.address2}"></span></p>
-    <p>배송 메세지: <span th:text="${order.shipMessage}"></span></p>
-    <p>결제 금액: <span th:text="${order.price}"></span></p>
-    <p>결제 방식:
-      <span th:switch="${order.payment}">
+
+<div layout:fragment="content">
+  <div class="container">
+    <h1>주문에 성공하였습니다. 주문 정보를 확인해주세요.</h1>
+    <div class="order-info" th:if="${order}">
+      <h2>주문 정보</h2>
+      <table class="table">
+        <thead>
+        <tr>
+          <th></th>
+          <th>제목</th>
+          <th>수량</th>
+          <th>한 권당 가격</th>
+        </tr>
+        </thead>
+        <tbody>
+        <tr th:each="orderItem : ${order.orderItems}">
+          <td>
+            <img th:src="${orderItem.book.imageUrl}" alt="책 이미지" width="100" height="150" />
+          </td>
+          <td th:text="${orderItem.book.title}"></td>
+          <td th:text="${orderItem.quantity}"></td>
+          <td th:text="${orderItem.book.price}"></td>
+        </tr>
+        </tbody>
+      </table>
+      <p>주문번호: <span th:text="${order.orderId}"></span></p>
+      <p>주문 날짜: <span th:text="${order.orderDate}"></span></p>
+      <p>이름: <span th:text="${order.name}"></span></p>
+      <p>전화번호: <span th:text="${order.phoneNumber}"></span></p>
+      <p>주소: <span th:text="${order.address.address1}"></span>
+        <span th:text="${order.address.address2}"></span></p>
+      <p>배송 메세지: <span th:text="${order.shipMessage}"></span></p>
+      <p>결제 금액: <span th:text="${order.price}"></span></p>
+      <p>결제 방식:
+        <span th:switch="${order.payment}">
         <span th:case="creditCard">신용 카드</span>
         <span th:case="bankTransfer">무통장 입금</span>
         <span th:case="mobile">휴대폰 결제</span>
         <span th:case="*" th:text="${order.payment}"></span>
       </span></p>
-    <p>카드 정보:
-      <span th:switch="${order.card}">
+      <p>카드 정보:
+        <span th:switch="${order.card}">
         <span th:case="shinhanCard">신한카드</span>
         <span th:case="KBCard">KB국민카드</span>
         <span th:case="samsungCard">삼성카드</span>
@@ -67,18 +111,19 @@
         <span th:case="NHCard">NH농협카드</span>
         <span th:case="*" th:text="${order.card}"></span>
       </span>
-    </p>
-    <div>
-      <p>주문 유형</p>
-      <span th:if="${order.orderType == 1}">일반책 구매</span>
+      </p>
+      <div class="order-type">
+        <p>주문 유형</p>
+        <span th:if="${order.orderType == 1}">일반책 구매</span>
+      </div>
+      <br/> <!-- 빈 줄 추가 -->
+      <div class="text-right"> <!-- 오른쪽 정렬 -->
+        <a href="/" class="btn btn-primary">메인으로 돌아가기</a>
+      </div>
     </div>
-    <br/> <!-- 빈 줄 추가 -->
-    <div class="text-right"> <!-- 오른쪽 정렬 -->
-      <a href="/main" class="btn btn-primary">메인으로 돌아가기</a>
+    <div th:if="${order == null}">
+      <h2>주문 정보가 없습니다.</h2>
     </div>
-  </div>
-  <div th:if="${order == null}">
-    <h2>주문 정보가 없습니다.</h2>
   </div>
 </div>
 </body>
