+++ src/main/resources/templates/cart/noCart.html
@@ -9,18 +9,32 @@
   <link href="https://fonts.googleapis.com/css?family=Nanum+Gothic" rel="stylesheet">
   <style>
     .container {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      height: 100vh;
-      flex-direction: column;
+        display: flex;
+        justify-content: center;
+        align-items: flex-start;
+        width: 1000px;
+        margin: 0 auto;
+        padding: 30px;
+    }
+
+    h3 {
+        display: block;
+        margin-bottom: 10px;
+    }
+
+    a {
+        display: block;
     }
   </style>
 </head>
 <body>
+
 <div layout:fragment="content">
-  <h1>장바구니가 존재하지 않습니다</h1>
-  <a href="/main">메인화면으로 돌아가기</a>
+  <div class="container">
+    <h3>장바구니가 존재하지 않습니다</h3>
+    <br>
+    <a href="/main">메인화면으로 돌아가기</a>
+  </div>
 </div>
 </body>
 </html>
